% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDistanceMatrix.R
\docType{methods}
\name{calcDistanceMatrix}
\alias{calcDistanceMatrix}
\title{Calculate Distance Matrix}
\usage{
\S4method{calcDistanceMatrix}{trModel}(object)
}
\arguments{
\item{object}{A \code{\linkS4class{trModel}} object}
}
\value{
The input \code{\linkS4class{trModel}} object where the \code{@distance} slot has been replaced by the
calculated distance matrix.
}
\description{
Produces a matrix of distance values where each row represents a receiver and each column represents a transmitter.
}
