% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTstart.R
\name{calcTstart}
\alias{calcTstart}
\title{Calculate Start Temp for simulated annealing}
\usage{
calcTstart(object, nswaps, targetProb)
}
\arguments{
\item{object}{The \code{\linkS4class{trModel}} object}

\item{nswaps}{How many random energies/swaps should be calculated.}

\item{targetProb}{The target probability desired for setting the initial temperature. \code{Tstart} calculated as \eqn{\sigma / ln(targetProb)}}
}
\description{
The goal of this function is to determine the appropriate starting temp so that the 'average' swap has a \code{targetProb} chance of
being accepted.  Obtained by simulating \code{nswaps} swaps and then calculating the temperature using
\eqn{\sigma / ln(targetProb) = Tstart}
}
