% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populateSpace.R
\docType{methods}
\name{populateSpace}
\alias{populateSpace}
\title{Populate Space with Transmitters and Receivers}
\usage{
\S4method{populateSpace}{trModel}(object, spaceRes = NULL)
}
\arguments{
\item{spaceRes}{The resolution of the space grid. The grid will be symmetric with \code{spaceRes} columns and \code{spaceRes} rows.
If provided, the returned \code{moodel} will have it's \code{@spaceRes} parameter replaced with the user provided parameter.}
}
\description{
Populates a grid that is \code{spaceResXspaceRes} with transmitters and receivers.  Transmitters are represented by a \code{1} and
receivers by a \code{-1}.
}
\details{
Currently uses sampling without replacement, may want to consider allowing replacement and handling collisions appropriately.
}
