% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomEnergy.R
\docType{methods}
\name{randomEnergy,trModel-method}
\alias{randomEnergy,trModel-method}
\title{Calculate a random energy/cost state for simmulated annealing}
\usage{
\S4method{randomEnergy}{trModel}(object)
}
\description{
Calculate a random energy/cost state for simmulated annealing
}
