% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simAnneal.R
\docType{methods}
\name{simAnneal,trModel,numeric,numeric-method}
\alias{simAnneal,trModel,numeric,numeric-method}
\title{Simulated Annealing}
\usage{
\S4method{simAnneal}{trModel,numeric,numeric}(object, tstart, tstop, alpha, n)
}
\arguments{
\item{object}{A \code{\linkS4class{trModel}} object}

\item{tstart}{The starting temperature}

\item{tstop}{The final temperature, once the temperature drops below this value the process will stop}

\item{alpha}{The multiplicitave temperature decrement}

\item{n}{The number of iterations per temperature before moving to the next temperature}
}
\description{
Simulated Annealing
}
