% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simAnneal.R
\docType{methods}
\name{simAnneal,trModel,numeric,numeric-method}
\alias{simAnneal,trModel,numeric,numeric-method}
\title{Simulated Annealing}
\usage{
\S4method{simAnneal}{trModel,numeric,numeric}(object, tstart, tstop, alpha, n,
  stability, endPlot = T, diagnostics = F, cluster = F)
}
\arguments{
\item{object}{A \code{\linkS4class{trModel}} object}

\item{tstart}{The starting temperature}

\item{tstop}{The final temperature, once the temperature drops below this value the process will stop}

\item{alpha}{The multiplicitave temperature decrement}

\item{n}{The number of iterations per temperature before moving to the next temperature}

\item{stability}{The number of temperature increments with the same end probability that causes early termination. If \code{0}
stability checking will not be performed.}
}
\description{
Plot info: Red line is optimal solution.  Green point is the best solution reached over the entire run, always plotted as the
last point regardless of when the best solution was actually reached.  Each point plotted is the probability at the end of a
given temperature state.
}
\details{
Stability based on best solution reached and checked after each temperature state.
}
